Create elbow tests.r
This code analyzes different setting to a data set clustering (K-means and Ward) based on a predefined data set header. It analyzes the data based on the elbow and generating the respective graphics for each analyzes method. in addition, the code also evaluates the clustering validity based on PBM and Ball and Hall methods.